Done with subtask5

+++ fund1/libcli3/libcli3/parsing/types.h
@@ -20,7 +20,7 @@ typedef struct {
   parsing_param_type_t param_type;
 
   union {
-    double fp;
+    long double fp;
     int integer;
     char string[PARSING_MAX_STRING_LENGTH];
   };
